requirements.txt
requests
zeep
Pillow
bs4